Conclusao da parte de Dtos e Data Annotations

+++ Back/src/CampEventos.API/Controllers/EventoController.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using CampEventos.Application.Contratos;
-using CampEventos.Domain;
-using CampEventos.Persistence.Contratos;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using CampEventos.Application.Dtos;
 
 namespace CampEventos.API.Controllers
 {
@@ -26,7 +23,7 @@ namespace CampEventos.API.Controllers
         try
         {
             var eventos = await _eventoService.GetAllEventosAsync(true);
-            if (eventos == null) return NotFound("Nenhum evento encontrado");
+            if (eventos == null) return NoContent();
         
             return Ok(eventos);
         }
@@ -45,7 +42,7 @@ namespace CampEventos.API.Controllers
         try
         {
             var evento = await _eventoService.GetEventoByIdAsync(id, true);
-            if (evento == null) return NotFound("Eventos por id não encontrado.");
+            if (evento == null) return NoContent();
         
             return Ok(evento);
         }
@@ -75,12 +72,12 @@ namespace CampEventos.API.Controllers
     }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
         try
         {
             var evento = await _eventoService.AddEventos(model);
-            if (evento == null) return BadRequest("Erro ao tentar adcionar evento.");
+            if (evento == null) return NoContent();
         
             return Ok(evento);
         }
@@ -93,12 +90,12 @@ namespace CampEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {    
         try
         {
             var evento = await _eventoService.UpdateEvento(id, model);
-            if (evento == null) return BadRequest("Eventos por tema não encontrado.");
+            if (evento == null) return NoContent();
         
             return Ok(evento);
         }
@@ -115,13 +112,15 @@ namespace CampEventos.API.Controllers
         {
         try
         {
+            var evento = await _eventoService.GetEventoByIdAsync(id, true);
+            if (evento == null) return NoContent();
+
             return await _eventoService.DeleteEvento(id) ? 
                     Ok("Deletado") : 
-                    BadRequest("Evento não deletado");
+                    throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento.");
         }
         catch (Exception ex)
         {
-            
            return this.StatusCode(StatusCodes.Status500InternalServerError,
            $"Erro ao tentar recuperar eventos. Erro: {ex.Message}");
         }
